task/EUAI-25-Admin-endpoints
Admins have complete CRUD access over ParkingSpaces in their Manager's Parking.
