--- src/main/java/com/chameleon/estaciona_uai/api/base_user/exception/BaseUserLoginInvalidCredentialsException.java
@@ -1,7 +1,0 @@
-package com.chameleon.estaciona_uai.api.base_user.exception;
-
-public class BaseUserLoginInvalidCredentialsException extends RuntimeException {
-    public BaseUserLoginInvalidCredentialsException(String email) {
-        super("User login attempt with invalid credentials for email '" + email + "'");
-    }
-}
