--- src/main/java/com/chameleon/estaciona_uai/api/customer/CustomerRepository.java
@@ -1,11 +1,0 @@
-package com.chameleon.estaciona_uai.api.customer;
-
-import com.chameleon.estaciona_uai.domain.user.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, UUID> {
-}
