+++ src/main/java/com/chameleon/estaciona_uai/api/user/admin/dto/AdminManagesParkingSpaceRequest.java
@@ -1,0 +1,8 @@
+package com.chameleon.estaciona_uai.api.user.admin.dto;
+
+import lombok.Data;
+
+@Data
+public class AdminManagesParkingSpaceRequest {
+    private String identifier;
+}
