+++ src/main/java/com/chameleon/estaciona_uai/api/user/base_user/dto/BaseUserLoginRequest.java
@@ -1,0 +1,17 @@
+package com.chameleon.estaciona_uai.api.user.base_user.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BaseUserLoginRequest {
+    @NotBlank(message = "Email is required")
+    @Email(message = "Email should be valid")
+    private String email;
+
+    @NotBlank(message = "Password is required")
+    private String password;
+}
