+++ src/main/java/com/chameleon/estaciona_uai/api/user/base_user/exception/UnexpectedBaseUserTypeException.java
@@ -1,0 +1,7 @@
+package com.chameleon.estaciona_uai.api.user.base_user.exception;
+
+public class UnexpectedBaseUserTypeException extends RuntimeException {
+    public UnexpectedBaseUserTypeException(String email) {
+        super("Unexpected UserType for email '" + email + "'.");
+    }
+}
