+++ src/main/java/com/chameleon/estaciona_uai/api/user/customer/CustomerRepository.java
@@ -1,0 +1,11 @@
+package com.chameleon.estaciona_uai.api.user.customer;
+
+import com.chameleon.estaciona_uai.domain.user.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, UUID> {
+}
