+++ src/main/java/com/chameleon/estaciona_uai/domain/parking/parking_space/ParkingSpace.java
@@ -1,12 +1,14 @@
 package com.chameleon.estaciona_uai.domain.parking.parking_space;
 
 import com.chameleon.estaciona_uai.domain.parking.Parking;
+import com.chameleon.estaciona_uai.domain.user.Admin;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Entity
@@ -16,15 +18,45 @@ import java.util.UUID;
 @NoArgsConstructor
 @Table(name = "parking_space")
 public class ParkingSpace {
+
     @Id
     @GeneratedValue
     private UUID id;
 
-    @ManyToOne
+    private String identifier;
+
+    @ManyToOne(optional = false)
     private Parking parking;
 
     @Enumerated(EnumType.STRING)
-    private ParkingSpaceStatus status;
+    private ParkingSpaceStatus status = ParkingSpaceStatus.FREE;
 
-    private String identifier;
+    private LocalDateTime deletedAt;
+
+    @Version
+    private long version;
+
+    /* ---------- domain helpers ---------- */
+
+    public void assertFree() {
+        if (isDeleted()) {
+            throw new IllegalStateException("DELETED");
+        }
+        if (status != ParkingSpaceStatus.FREE) {
+            throw new IllegalStateException("NOT_FREE");
+        }
+    }
+
+    public void validateOwnership(Admin admin) {
+        if (isDeleted()) {
+            throw new IllegalStateException("DELETED");
+        }
+        if (!parking.getManager().equals(admin.getManager())) {
+            throw new SecurityException("NOT_OWNER");
+        }
+    }
+
+    public void softDelete() { this.deletedAt = LocalDateTime.now(); }
+
+    private boolean isDeleted() { return deletedAt != null; }
 }
